Add HomeWorks queries. Need UnitTesting

+++ src/main/java/com/goit/javadev/tables/manytomany/ManyToMany.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public interface ManyToMany<T> {
     boolean insertKey(T element);
-    boolean deleteKey(int key1, int key2);
+    boolean deleteKey(long key1, long key2);
     Optional<List<T>> getAll();
     int createKeysFromList(List<T> list);
     int insertKeysFromJsonFile(String jsonFilePath);
